requirements.txt
python=3.9.17
joblib=1.4.2
matplotlib=3.8.4
openpyxl=3.1.5
pandas=1.4.4
scikit-learn=1.2.2
scipy=1.12.0
h5py==3.11.0
keras==2.15.0
numpy==1.26.4
tensorflow==2.15.1